auth token hashes are now stored in the DB, and cached upon retrieval
