+++ server/data/AuthDAO/AuthToken.js
@@ -1,0 +1,8 @@
+class AuthToken {
+    constructor(token, tokenHash) {
+        this.token = token;
+        this.tokenHash = tokenHash;
+    }
+}
+
+module.exports = AuthToken;
