+++ server/mediaUtil/gifMaker.js
@@ -2,7 +2,7 @@ const { spawn } = require('child_process');
 const fs = require('fs');
 const path = require('path');
 const { FilePaths } = require('../const');
-const Job = require('../uploadAPI/Job')
+const Job = require('../api/uploadAPI/Job')
 
 
 const QUALITY = {
